test: reorganize test folders to match commands
